Move the subscription purchase wide event to the subscription library
